+++ Sources/StreamChat/Database/DTOs/MessageDTO.swift
@@ -127,8 +127,6 @@ class MessageDTO: NSManagedObject {
     @NSManaged var defaultSortingKey: DBDate!
     
     @NSManaged var channelRole: String?
-    @NSManaged var memberNotificationsMuted: Bool
-    @NSManaged var memberExtraData: Data?
     @NSManaged var deletedForMe: Bool
 
     override func willSave() {
@@ -940,20 +938,8 @@ extension NSManagedObjectContext: MessageDatabaseSession {
         dto.parentMessageId = payload.parentId
         dto.showReplyInChannel = payload.showReplyInChannel
         dto.replyCount = Int32(payload.replyCount)
-        if let member = payload.member {
-            if let role = member.channelRole?.rawValue {
-                dto.channelRole = role
-            }
-            dto.memberNotificationsMuted = member.notificationsMuted
-            do {
-                dto.memberExtraData = try JSONEncoder.default.encode(member.extraData)
-            } catch {
-                log.error(
-                    "Failed to encode member extra payload for Message with id: <\(dto.id)>, using default value instead. "
-                        + "Error: \(error)"
-                )
-                dto.memberExtraData = nil
-            }
+        if let role = payload.member?.channelRole?.rawValue {
+            dto.channelRole = role
         }
 
         do {
@@ -1853,25 +1839,10 @@ private extension ChatMessage {
 
         let readBy = Set(dto.reads.compactMap { try? $0.user.asModel() })
         
-        let member: ChatMessage.MemberInfo? = {
-            guard dto.channelRole != nil || dto.memberExtraData != nil else { return nil }
-
-            let extraData: [String: RawJSON]
-            do {
-                extraData = try JSONDecoder.stream.decodeRawJSON(from: dto.memberExtraData)
-            } catch {
-                log.error(
-                    "Failed to decode member extra data for Message with id: <\(dto.id)>, using default value instead. Error: \(error)"
-                )
-                extraData = [:]
-            }
-
-            return ChatMessage.MemberInfo(
-                channelRole: dto.channelRole.map(MemberRole.init(rawValue:)),
-                notificationsMuted: dto.memberNotificationsMuted,
-                extraData: extraData
-            )
-        }()
+        var channelRole: MemberRole?
+        if let role = dto.channelRole {
+            channelRole = MemberRole(rawValue: role)
+        }
 
         let message = ChatMessage(
             id: id,
@@ -1924,7 +1895,7 @@ private extension ChatMessage {
                 updatedAt: $0.updatedAt.bridgeDate
             ) },
             sharedLocation: location,
-            member: member
+            channelRole: channelRole
         )
 
         if let transformer = chatClientConfig?.modelsTransformer {